requirements.txt
Appium-Python-Client>=3.0.0
Selenium==4.13.0
pytest>=7.4.2
allure-pytest>=2.13.2
selene>=2.0.0rc4
python-dotenv~=1.0.0
pydantic-settings>=2.0.3
requests~=2.31.0